fix(infrastructure): address CodeRabbit review issues
- Call ApplyGlobalQueryFilters in MystiraAppDbContext.OnModelCreating
- Fix AddPolyglotPersistenceWithDualWrite to properly use type parameters
- Add new AddPolyglotEntityWithDualWrite method for dual-write entity registration
- Fix AxisScores type from float to int to match domain model
- Delete unused local IUnitOfWork interface (use Application.Ports.Data)
- Add configureOptions parameter to Teams AddTeamsBotKeyed method

+++ packages/infrastructure/Mystira.Infrastructure.Data/PostgresDbContext.cs
@@ -290,8 +290,8 @@ public class PostgresDbContext : DbContext
                 .HasColumnType("jsonb")
                 .HasConversion(
                     v => JsonSerializer.Serialize(v, (JsonSerializerOptions?)null),
-                    v => JsonSerializer.Deserialize<Dictionary<string, float>>(v, (JsonSerializerOptions?)null) ?? new Dictionary<string, float>(StringComparer.OrdinalIgnoreCase))
-                .Metadata.SetValueComparer(CreateDictionaryComparer<string, float>());
+                    v => JsonSerializer.Deserialize<Dictionary<string, int>>(v, (JsonSerializerOptions?)null) ?? new Dictionary<string, int>(StringComparer.OrdinalIgnoreCase))
+                .Metadata.SetValueComparer(CreateDictionaryComparer<string, int>());
 
             // Indexes
             entity.HasIndex(e => e.ProfileId).HasDatabaseName("ix_player_scenario_scores_profile_id");
